faz os requisitos da primeira tela de usuario faltando a questao do multer, e faz integraçao com login

+++ back-end/src/database/migrations/20220221212203-create-product.js
@@ -12,7 +12,7 @@ module.exports = {
         type: Sequelize.STRING
       },
       price: {
-        type: Sequelize.DOUBLE
+        type: Sequelize.DECIMAL(10,2)
       },
       urlImage: {
         type: Sequelize.STRING,
